+++ lib/src/doc_scan_kit_method_channel.dart
@@ -24,7 +24,11 @@ class MethodChannelDocScanKit extends DocScanKitPlatform {
             ?.map(
           (e) {
             e as Map;
-            return ScanResult(imagePath: e['path'], imagesBytes: e['bytes']);
+            return ScanResult(
+                imagePath: e['path'],
+                imagesBytes: e['bytes'],
+                text: e['text'],
+                barcode: e['barcode']);
           },
         ).toList() ??
         [];