+++ lib/src/options/scan_result.dart
@@ -3,8 +3,12 @@ import 'dart:typed_data';
 class ScanResult {
   String? imagePath;
   Uint8List imagesBytes;
+  String? text;
+  String? barcode;
   ScanResult({
     required this.imagePath,
     required this.imagesBytes,
+    required this.text,
+    required this.barcode,
   });
 }